+++ landing/src/pages/chains/EthereumPage.vue
@@ -311,6 +311,13 @@
                             <span class="related-desc">Block explorer for Ethereum's primary testnet</span>
                         </div>
                     </router-link>
+                    <router-link to="/etherscan-alternative" class="related-link">
+                        <v-icon size="18" style="color: #5DAAE0;">mdi-compare-horizontal</v-icon>
+                        <div>
+                            <span class="related-title">Etherscan Alternative</span>
+                            <span class="related-desc">How Ethernal compares to Etherscan for custom EVM chains</span>
+                        </div>
+                    </router-link>
                     <router-link to="/pricing" class="related-link">
                         <v-icon size="18" style="color: #5DAAE0;">mdi-tag-outline</v-icon>
                         <div>